[fix] 몬스터 체력 int -> float로 수정

+++ Assets/_Scripts/Monster/Boss/BoneReaper/BoneReaper.cs
@@ -122,7 +122,7 @@ public class BoneReaper : BossMonster
         return NodeStatus.Failure;
     }
 
-    public void TakeDamage(int damage)
+    public void TakeDamage(float damage)
     {
         if (_dataHandler.CurHp <= 0)
             return;
