+++ Assets/_Scripts/Monster/Boss/BoneReaper/BoneReaperHead.cs
@@ -88,8 +88,8 @@ public class BoneReaperHead : MonoBehaviour, IDamageable
                 targetRb.AddForce(knockBackDir * 400);
             }
 
-            target.GetComponent<IDamageable>()?.TakeDamage((int)(_owner.MonsterData.AttackPower));
-            Debug.Log((int)(_owner.MonsterData.AttackPower));
+            target.GetComponent<IDamageable>()?.TakeDamage(_owner.MonsterData.AttackPower);
+            Debug.Log((_owner.MonsterData.AttackPower));
         }
     }
 
@@ -129,7 +129,7 @@ public class BoneReaperHead : MonoBehaviour, IDamageable
     }
 
     // 기타 메서드
-    public void TakeDamage(int damage)
+    public void TakeDamage(float damage)
     {
         if (_owner.MonsterData.CurHp <= 0 || _owner.IsOneFrameInvincible)
             return;
