+++ Assets/_Scripts/Monster/Core/Slime.cs
@@ -35,7 +35,7 @@ public class Slime : PatrolStateMonster
             targetRb.AddForce(knockBackDir * 400);
 
             IDamageable damageable = collision.GetComponent<IDamageable>();
-            damageable?.TakeDamage((int)_dataHandler.AttackPower);
+            damageable?.TakeDamage(_dataHandler.AttackPower);
         }
     }
 }
